third_party/blink/web_tests/wpt_internal/fenced_frame/urn.https.html

<!DOCTYPE html>
<title>Fenced frames urn mapping</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<script src="/common/utils.js"></script>

<body>
<script>
promise_test(async (t) => {
  const key = token();

  attachFencedFrame(await generateURN("resources/embeddee.html", [key]), 'opaque-ads');
  const result = await nextValueFromServer(key);

  assert_equals(result, "PASS",
                "The fenced frame with src=urn:uuid should load");
}, 'urn:uuid Loading Test');

</script>
</body>
</html>
